+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- stored: fix race condition [PR #2362]
+
 ## [24.0.5] - 2025-09-02
 
 ### Documentation
@@ -511,4 +514,5 @@ It is therefore strongly suggested to immediately schedule a full backup of your
 [PR #2343]: https://github.com/bareos/bareos/pull/2343
 [PR #2345]: https://github.com/bareos/bareos/pull/2345
 [PR #2351]: https://github.com/bareos/bareos/pull/2351
+[PR #2362]: https://github.com/bareos/bareos/pull/2362
 [unreleased]: https://github.com/bareos/bareos/tree/master